lab12/pyproject.toml
[project]
name = "MLOps_course_AGH_lab_02"
version = "1.0.0"
description = ""
readme = "README.md"

requires-python = "==3.11.*"

dependencies = [
    "apache-airflow[celery,pandas,standard,uv]==3.1.5",
    "cloudpickle",
    "pandas",
    "requests",
    "twelvedata",
    "pyarrow"
]

[dependency-groups]
dev = [
    "ruff"
]

airflow_common = [
    "apache-airflow[celery,postgres,standard,uv]==3.1.5",
    "requests",
]